qgepplugin/interlis/requirements.txt
sqlalchemy~=1.3.22
geoalchemy2~=0.8.4

# dev only
jinja
